more UX and errorMsg changes

+++ src/components/loginpage/sections/LoginForm.jsx
@@ -1,16 +1,31 @@
-import { useEffect, useReducer } from "react";
+import { useReducer } from "react";
 import { useState } from "react";
 import PrimaryButton from "../../atoms/buttons/PrimaryButton";
 import EmailField from "../../atoms/forms/EmailField";
 import PasswordField from "../../atoms/forms/PasswordField";
 import styles from "../../shared/Forms.module.css";
+import UnauthorisedModal from "../../atoms/posts/UnauthorisedModal";
+import Modal from "react-modal";
+
+const customStyles = {
+  content: {
+    top: "50%",
+    left: "50%",
+    right: "auto",
+    bottom: "auto",
+    marginRight: "-50%",
+    transform: "translate(-50%, -50%)",
+    border: 0,
+  },
+};
 
 export default function LoginForm({ isLoggedIn, setIsLoggedIn }) {
   const [valid, setValid] = useState(undefined);
-  const [error, setError] = useState("");
   const [errorEmail, setErrorEmail] = useState("");
   const [errorPassword, setErrorPassword] = useState("");
   const [emailAvailable, setEmailAvailable] = useState("");
+  const [isOpen, setIsOpen] = useState(false);
+  const [errorMsg, setErrorMsg] = useState("");
 
   const reducer = (state, newValues) => {
     return { ...state, ...newValues };
@@ -28,15 +43,16 @@ export default function LoginForm({ isLoggedIn, setIsLoggedIn }) {
     });
   };
   //Function to verify the email and password inputs
-  const verifyInputs = () => {
+  const verifyInputs = (event) => {
+    event.preventDefault();
     if (formValues.email === "" || formValues.password === "") {
       setValid(false);
-      setError("");
+      setErrorMsg(
+        `All fields marked with a '*' have to be filled out before submitting!`
+      );
     } else {
       setValid(true);
-      /* // Then we call the backend and verify the user.
-      		const resultFromServer = "Invalid Login";
-      		setError(resultFromServer); */
+      setErrorMsg("");
       loginUser();
     }
   };
@@ -90,31 +106,35 @@ export default function LoginForm({ isLoggedIn, setIsLoggedIn }) {
     fetch("http://localhost:3004/auth/login", options)
       .then((response) => response.json())
       .then((response) => {
-        dispatch({
-          ["email"]: "",
-          ["password"]: "",
-        });
-        localStorage.clear();
-        localStorage.setItem("token", response.access_token);
-        localStorage.setItem("user", JSON.stringify(response.user));
-        setTimeout(() => localStorage.clear(), 3600000);
-        setIsLoggedIn(true);
         console.log(response);
+        if (response.statusCode === 401) {
+          setErrorMsg(
+            "Ops! Something went wrong. Please try to log in again with the correct email and password."
+          );
+          setIsOpen(true);
+        } else {
+          dispatch({
+            ["email"]: "",
+            ["password"]: "",
+          });
+          localStorage.clear();
+          localStorage.setItem("token", response.access_token);
+          localStorage.setItem("user", JSON.stringify(response.user));
+          setTimeout(() => localStorage.clear(), 3600000);
+          setIsLoggedIn(true);
+          setErrorMsg("You are now successfully logged in!");
+          setIsOpen(true);
+        }
       })
       .catch((err) => {
         console.error(err);
-        checkEmail();
-        validatePassword();
       });
-    console.log(setIsLoggedIn());
   }
 
   return (
-    <section className={styles.formWrapper}>
+    <section onSubmit={verifyInputs} className={styles.formWrapper}>
       <h1>Log In</h1>
       <form className={styles.form}>
-        <p>{error}</p>
-
         <EmailField
           value={formValues.email}
           onChange={updateFormValue}
@@ -132,16 +152,29 @@ export default function LoginForm({ isLoggedIn, setIsLoggedIn }) {
         />
 
         <PrimaryButton
-          type="button"
+          type="submit"
           id="submit"
-          onClick={verifyInputs}
           text="Submit"
           errorEmail={errorEmail}
         />
 
-        {valid && <p>Login successful!</p>}
-        {valid === false && <p>Login failed</p>}
+        {!valid && <p>{errorMsg}</p>}
       </form>
+      <Modal
+        isOpen={isOpen}
+        onRequestClose={() => setIsOpen(false)}
+        contentLabel="Example Modal"
+        style={customStyles}
+        shouldCloseOnOverlayClick
+      >
+        <UnauthorisedModal
+          style={styles}
+          onClick={() => setIsOpen(false)}
+          errorMsg={errorMsg}
+          isLoggedIn={isLoggedIn}
+          title="Welcome!"
+        ></UnauthorisedModal>
+      </Modal>
     </section>
   );
 }
